Update ordered-set agg preprocess step for skew
Prior to this commit, the preprocessing for supported ordered-set agg
would split the ordered-set agg into a NLJ between total_count and the
CTEConsumer for the input table with a gp_percentile_* GbAgg on top.
For skewed dataset, this wouldn't be as performant as the JOIN would
return all the rows. This commit updates the code to split the
ordered-set agg into a NLJ between CTEConsumer for deduplicated data and
total_count on that CTE with a gp_percentile_* GbAgg on top. Since, we
deduplicate the data, we also pass along the count of each distinct row
as peer_count to gp_percentile_* agg.
Below, is the preporcessed query:
Input query:
```
+--CLogicalGbAgg( Global )
   |--CLogicalGet "t" ("t")
   +--CScalarProjectList
      +--CScalarProjectElement "percentile_cont"
         +--CScalarAggFunc (percentile_cont , Distinct: false , Aggregate Stage: Global)
            |--CScalarValuesList
            |  +--CScalarIdent "a" (0)
            |--CScalarValuesList
            |  +--CScalarConst (0.250)
            |--CScalarValuesList
            |  +--CScalarSortGroupClause(tleSortGroupRef:0,eqop:96,sortop:97,nulls_first:false,hashable:true)
            +--CScalarValuesList
```

Output preprocessed query:
```
Common Table Expressions:
+--CLogicalCTEProducer (0)
   +--CLogicalGbAgg( Global ) Grp Cols: ["a" (10)]
      |--CLogicalGet "t" ("t")
      +--CScalarProjectList
         +--CScalarProjectElement "ColRef_0009" (11)
            +--CScalarAggFunc (count , Distinct: false , Aggregate Stage: Global)
               |--CScalarValuesList
               |  +--CScalarIdent "a" (10)
               |--CScalarValuesList
               |--CScalarValuesList
               +--CScalarValuesList

Algebrized preprocessed query:
+--CLogicalCTEAnchor (0)
   +--CLogicalGbAgg( Global )
      |--CLogicalLimit ( (97,1.0), "a" (0), NULLsLast )  global
      |  |--CLogicalNAryJoin
      |  |  |--CLogicalCTEConsumer (0), Columns: ["a" (0), "ColRef_0009" (9)]
      |  |  |--CLogicalProject
      |  |  |  |--CLogicalGbAgg( Global )
      |  |  |  |  |--CLogicalCTEConsumer (0), Columns: ["a" (19), "ColRef_0009" (20)]
      |  |  |  |  +--CScalarProjectList
      |  |  |  |     +--CScalarProjectElement "ColRef_0035" (35)
      |  |  |  |        +--CScalarAggFunc (sum , Distinct: false , Aggregate Stage: Global)
      |  |  |  |           |--CScalarValuesList
      |  |  |  |           |  +--CScalarIdent "ColRef_0009" (20)
      |  |  |  |           |--CScalarValuesList
      |  |  |  |           |--CScalarValuesList
      |  |  |  |           +--CScalarValuesList
      |  |  |  +--CScalarProjectList
      |  |  |     +--CScalarProjectElement "ColRef_0036" (36)
      |  |  |        +--CScalarFunc (int8)
      |  |  |           +--CScalarIdent "ColRef_0035" (35)
      |  |  +--CScalarConst (1)
      |  |--CScalarConst (0)
      |  +--CScalarConst (null)
      +--CScalarProjectList
         +--CScalarProjectElement "percentile_disc" (8)
            +--CScalarAggFunc (percentile_disc , Distinct: false , Aggregate Stage: Global)
               |--CScalarValuesList
               |  |--CScalarIdent "a" (0)
               |  |--CScalarConst (0.250)
               |  |--CScalarIdent "ColRef_0036" (36)
               |  +--CScalarIdent "ColRef_0009" (9)
               |--CScalarValuesList
               |--CScalarValuesList
               +--CScalarValuesList
```

This also inclodes, updating C function for gp_percentile

Since we pass the peer_count value along with the total_count, we
need to consider it while calculating percentile values.
This commit also updates the transition functions to non-strict as for
strict transition function `advance_aggregates()` calls
`ExecInterpExpr()` which initializes the transition value for the first
row in the group as part of `EEOP_AGG_INIT_TRANS` step. This results in
calling the transition function from second row with first row passed in
as the previos state value. This worked fine previously as we read all
rows and the peer_count would always be 1, but now since we need to read
the information for peer_count for each row, initializing the first row
doesn't work.
Since, the transition function isn't strict anymore, we
explicitly handle for NULL inputs.