Remove WITH OIDS from ORCA code base (#13920)
Postgres commit 578b229 (from Postgres 12 merge) removed WITH OIDS
support. That eliminated the "specialness" of oid columns which
previously were not stored as a normal column, but as part of the tuple
header. Now, in pg_class for example, it is a normal column.

ORCA had a framework in place to handle this "specialness".  During the
Postgres 12 merge the framework was kept in place and hardcoded to false
with a FIXME to remove later. This commit does that.