Address FIXMEs in gpdbwrappers.cpp file (#14104)
* Remove FIXME for signature change

As there are other functions that also take in Node as an input, its
better to keep it as is instead of changing the signature and breaking
something not captured as ICW.

* Remove dead code

This commit was to address the FIXME that removes CountLeafPartTables().
On looking at the call hierarchy CountLeafPartTables() was called by
RetrieveNumChildPartitions() -> GenerateStatsForSystemCols() ->
RetrieveColStats() for attno < 0 (system columns). Since we do not
extract/use the stats on system columns, the entire call stack is not
used code. This commit removes this part of the code altogether.

It also previously called RelPartIsNone() call to which was removed,
thus removing this function too.

This commit also removes FIXME for collation.