Store aggregate replication safety info in catalog
In GPDB certain aggregates (specifically order agnostic functions) can
be safely executed on replicated slices. The alternative is to execute
the aggregate on a single segment and then broadcast to the others. The
performance impact of broadcast can be significant on large replicated
datasets.

This commit adds an entry in pg_aggregate to distinguish between
aggregates that can and cannot be safely executed on replicated slices.
By default aggregates are not considered safe to execute on replicated
slices. A user can create safe aggregates by explicitly specifying a new
optional parameter 'repsafe' during CREATE AGGREGATE command.