Fix bug of wrong idx position in targetlist as ExecTupleSplit (#14954)
Targetlists are allowed to project non-group columns which not wrapped in Aggregate funciton
when group-by clause includes any contraint primary key, such as:
```
select count(distinct a), count(distinct b), e from t1 group by a, b;
if (a, b) is a primary of table t1, column e could be projected directly.
```
As case above, non-group columns are not only in final targetlist not also in targetlist 
of any subplans, such as node TupleSplit. 
The commit is going to fix inconsistant relative position of those columns in Node TupleSplit targetlist.
We could store relative postion of column towards targetlist(te->resno) from one instead of zero
in the whole life time of ExecTuplesplit.