Update CDXLScalarArrayCoerceExpr to pass elemexpr
As part of GP6 and older versions, ArrayCoerceExpr struct held the
information of the cast function as function oid and a boolean
isExplicit for representing if its an explicit or implicit cast.
Starting PG11 this was changed to an Expr elemexpr.
This commit updates the representation of CDXLScalarArrayCoerceExpr in
ORCA so that we can correctly roundtrip the Expr *elemexpr.
Thus addressing the 2 FIXME's that are removed as part of this commit.