PAX: Add exception catch for some C++ callback
Since CBDB is written in C, PAX is written in C++.

In the AM interface, if PAX generates an exception but does not catch it,
the exception will cause the CBDB process to abort.

Added CBDB_TRY/CBDB_CATCH_DEFAULT/CBDB_END_TRY to convert exceptions into
long jumps so that exceptions thrown by PAX can be catch and cleaned up.