PAX: Tuple will in-place updated, so should not keep the memory ptr
In order to reduce the cost of tupleslot copying, `OrcWriter` will always
keep the buffer pointer in the `TupleSlot` but not copy.

But for non-fixed-length columns, tupleslots are always reused when writing,
which means that previous tupleslots will be overwritten by subsequent
tupleslots. So in PAX, we can only copy tupleslot to memory.

consider replace `std::vector<DataBuffer<char *>>` with `DataBuffer<char *>`,
then no need copy it twice.