Fix memory problem caused by DatumFromCString
When PAX insert a non-fixed column, it will remove the VARHDR of datum
and insert it into the file. When reading tuple, PAX have to combine it
with `DatumFromCString`.

`DatumFromCString` is used to add the 4B (VARHDRSZ) header to the datum.
But the memory in `DatumFromCString` is reallocated. This causes read
performance to be affected.

In the current changes, PAX no longer removes VARHDR, but directly stores
and reading the datum.