Query with SPI will call the ExecutorEnd multiple times in recursive style
Query with SPI will call the ExecutorEnd multiple times in recursive style,
which causes the shared memory to be released by mistake on gp_writer, and
results in null shared memory when executing the scan.

Temporary solution: if ExecutorEnd is called because of the end of a
function, skip releasing the shared memory.