Fix: byte alignment is necessary when writing data for non-fixed columns
when we write to variable-length types which typlen is -1, we need to maintain byte alignment(MAXALIGN is used for default value）. If we use a compact layout, this will cause errors in heap_compute_data_size function when calculating data_length.

And we don't need to fill the padding with 0. the length of the header is parsed througth VARSIZE_ANY, which is unrelated to whether the tail is '\0' or not.

```
create extension pax;
set default_table_access_method='pax';
CREATE TABLE brinopers (colname name, typ text,
	op text[], value text[], matches int[],
	check (cardinality(op) = cardinality(value)),
	check (cardinality(op) = cardinality(matches)));
INSERT INTO brinopers VALUES
	('int4rangecol', 'int4range',
	 '{<<, &<, &&, &>, >>, @>, <@, =, <, <=, >, >=}',
	 '{"[10000,)","[10000,)","(,]","[3,4)","[36,44)","(1500,1501]","[3,4)","[222,1222)","[36,44)","[43,1043)","[367,4466)","[519,)"}',
	 '{53, 52, 51, 54, 50, 22, 72, 1, 74, 75, 34, 21}');
SELECT colname, oper, typ, value[ordinality], matches[ordinality] FROM brinopers, unnest(op) WITH ORDINALITY AS oper order by colname, typ;
```