Always use the database's default tablespace for the auxiliary relation of PAX table
The tablespace of the auxiliary relation of the PAX table is changed from following the main table to the database's default tablespace. The catalog resides in the database's default tablespace, not always pg_default if the database uses tablespace other than pg_default. It's better to save the auxiliary relations like the builtin catalog.

This avoids undetermined behavior of the tablespace for PAX.