Refactor string length checks in Orca
Previously, we would use GPOS_WSZ_LENGTH, which is just a wrapper around
gpos::clib::Wcslen(x), which checks the length by iterating over the
string. We already store the string length within the string, so we
should use that instead. This shaves off ~8% of optimization time in my
tests when the relcache is populated.