Introducing Non-Replicated Distribution Spec
**Motivation**

Currently, we don't support DPE when joining a partition table with data
universally available. Such data could be an unnested array or a TVF. To allow
DPE, hash join places the partition table as the outer child, and the partition
selector as the inner child. A common scenario is where the partition table is
randomly distributed on the join key (partition key is usually not the
distribution key), and the non-partitioned table is broadcasted. However, if the
inner child is universal, we enforce the outer child into a singleton to avoid
duplicates. Forcing a partition table into a singleton means gathering all the
data into the coordinator. As a result, the join order for DPE to happen is not
further explored.

Apparently, enforcing the outer table into a singleton when the inner side is
universal is overly conservative. `Universal` distribution spec is very much
like `Replicated`, and it should be joinable with `Non-Singleton` as well. This
PR relaxes the enforcement. This way, we avoid unnecessary Gather or duplicate
sensitive motions and generate better plans (including propagating DPE).

**Implementation**

`Non-Singleton` distribution spec is satisfiable by both `Hash` and `Random`. As
the name suggests, `Non-Singleton` excludes `Singleton`. Since the idea is to
allow `Universal` to be joinable with `Non-Singleton` and `Singleton`, we
introduce a new distribution spec `Non-Replicated`, satisfiable by `Hash`,
`Random`, and `Singleton`. This PR applies the `Non-Replicated` spec to hash
join and inner nested loop join.

We use `Non-Replicated` for two types of hash join optimization requests: (1)
hash distribute both sides, and (2) broadcast the inner side. In the first type
of optimization, if hash distributing both sides isn't doable (eg. columns not
hashable), we request a matching spec for one child given the other's
distribution spec. In the second type of optimization, we request the inner
child to be replicated. In both cases, if one child is `Universal`, we request
the other child to be `Non-Replicated` only if there's no duplicate concern.
Otherwise, we still request a `Singleton`.

Similarly in inner nested loop join, we request `Non-Replicated` for one child
if the other is `Universal`. Since inner join doesn't generate duplicates as far
as one child is deduplicated, it's safe to always pair `Non-Replicated` with
`Universal`.

The following two examples illustrates the "duplicate concern".

**_Eg.1_**
Inner join of `Universal` and `Non-Replicated` doesn't generate duplicates,
because the `Non-Replicated` side deduplicates the join output.

**_Eg.2_**
Left outer join of `Universal` and `Non-Replicated` generates duplicates,
because it outputs all the tuples from the outer table. Therefore, we request
the inner side to be a `Singleton` if the `Universal` table is on the outer side
of the left outer join.

**Discussions**

Distribution specs can be best understood by their enforcers. For example,
`Random` and `Non-Singleton` have very similar enforcers. Therefore, even though
`Non-Singleton` can be satisfied by `Random`, `Hash`, and even `Replicated` by
default, it bears more similarity to `Random` than to the other two specs.
Interestingly, we inherit `Hash` from `Random`, possibly because they share some
features such as handling duplicates. But since they have completely different
enforces, they should really be perceived as very different specs.

We build `Non-Replicated` as an inherited class of `Singleton` so it shares the
same enforcer -- Gather. There are two reasons for this implementation.

1. As the naming suggests, `Non-Replicated` is anything but `Replicated`. What
this means is `Non-Replicated` doesn't have duplicates. Correspondingly, its
enforcer's job is to deduplicate any given data distribution. Both `Hash` and
`Random` enforcers have mechanisms to handle duplicates by using a filter that's
not actually a physical motion. But that mechanism is more of an add-on feature
in case the given spec is `Universal`. The enforcers' main jobs are to hash
distribute or randomize the data, which would be an overkill for our purpose.
Comparatively, `Singleton` enforcer actually gets rid of the duplicates by
gathering the data into the coordinator.

2. `Non-Replicated` is introduced to replace existing usages of `Singleton` when
appropriate. If the given distribution spec doesn't satisfy what we request, it
would be ideal for the updated logic to behave like the original.
`Non-Replicated` offers more leeway to avoid enforcement. But in case we have to
enforce, `Non-Replicated` being an inheritance of `Singleton` will likely get us
the same plan as before.

**Tests**
This PR updated a lot of ORCA plans. Here are a few catagories --

1. Partition selector propagated to enable DPE
2. Gather motion enforced after join to benefit from MPP
3. Join order of nested loop join swapped as a result of added alternatives
4. Avoid redundant (non)physical motions