Fix: fs test will got some permissions errors
In some test cases of PAX Gtest a fixed path under /tmp is used, it may cause some
file permission issues.

The current change changes it to a relative path (relative to the execution path).

In addition, some test cases will not delete the test files during the teardown stage,
which will cause a problem when the test is repeatedly executed.