[ORCA] Allow index only scan on more index types (#16260)
In commit 3b72df18104, ORCA introduced support for index-only scan on
only b-tree indexes. That restriction was intended to reduce the scope
of the commit. This commit expands the index types that ORCA allows for
index-only scan by checking the index's access method for returnable
columns. Also updates MDP indexes with ReturnableColumns attribute.