Fixing column width of partitioned tables (#16282)
* Fixing column width of partitioned tables

In 6X while retrieving the column widths of a partitioned table, inherited
stats i.e. stats that cover all the child tables are used. This is done by
using method gpdb::GetAttStats. But in 7X the inherited stats are not used. In
7X the method gpdb::GetAttAvgWidth is getting used which generates stats for
individual tables and not for inheritance trees. This PR makes the behaviour of
7X similar to 6X by using inherited stats for partitioned tables.