Optimize string creation in Orca (#16332)
Previously, in order convert from a string literal into a CName (which
is heavily used in Orca), it would build this using helper methods that
created unnecessary data structures like DXL objects and dynamic
strings. Instead, build this in a simpler way. This reduces optimization
time by for queries ~12% when reading in strings to the relcache.