Relax typid check for empty data file
Data file may be empty in our current implementation. In this case,
the micro partition stats is also initialized, but the typid is not set yet.
The typid assert should only happen if the typid is set when inserting
the first tuple.

We relax this constraint to allow empty data file now.