Introduce: cpp-stub to mock global and private method
PAX already supports the gmock(google mock), but gmock does not support mocking
private functions or changing C function pointers.

In the unit test of PAX, some CBDB functions always need to be called, but it is
difficult for this kind of CBDB function to construct the returns required in the test case.

So PAX introduces a new library named `cpp-stub`. cpp-stub can replace some global
C functions or private functions with PAX implementations, making test cases easier to implement.