Save and validate opfamily instead of oids of compare functions
Previously, we save the oids of compare functions to validate
whether the comparison is meaningful. However the oids of
the compare functions varies for different data types.
This commit save the opfamily for validation. We assume that
the compare functions are always consistent as long as the opfamily
is the same.