Fix ORCA returning wrong column type modifier info (#16810)
Issue:
When executing a query with a false filter that results in zero rows
from a table, ORCA returns incorrect column type information in the plan.
The type modifier details for columns in the table metadata are absent.

RCA:
In 7X, redundant Result node in a planned_stmt is being removed
by function remove_redundant_results in orca.c.
Within remove_redundant_results, the outer result node's var expr
is being replaced with child result's const expr.
The const expr representing a column(tle) in a child Result node
of DXL or planned_stmt is created with default value of consttypmod.

So the replacement results in an erroneous consttypmod
for the column type information in the plan.

Fix:
To rectify this issue, the const.consttypmod value is now corrected
before the replacement of var with const takes place.

Regression test changes:
* Added a new test case with the same query
  used by external libraries for fetching table schema.
* Fixed the expected output for a similar case.