[ORCA] Use standard library for compile time assert (#16938)
In C++11 static_assert was introduced in the standard library. That
interface allows us to add useful error messages for when the assert
fails. Unfortunately, error message argument is required until C++17, so
existing callers need to be updated. Luckily there weren't many.

The existing implementation was likely derived from Van Horn 1997
Compile-time assertions in C++ [1]. Besides not supporting error
messages, it also renders unused variable warnings if used in the scope
of a function.

[1] https://dl.acm.org/doi/10.5555/282637.282656