Feature: use stripe stats instead of repeated mix/max update operations
When pax writes data, two kinds of statistical information level will be generated.
One is file-level statistical information, and the other is group-level statistical
information.

But after in WriteTuple, both levels of statistics are updated. In fact,
we only need to update the group-level statistics, then file-level statistics
can merge it.