Pax: Remove the struct CTupleSlot
CTupleSlot was originally used to encapsulate some operations on TupleTableSlot.

In fact, this is an additional object introduced. There are already enough functions in
PG to modify TupleTableSlot.

In the current version, this object will not only generate additional memory, but also
make the interface more complex. So the current change decide to remove it.