Fix for finding child output columns when parent is union while join pruning (#16960)
Fix issue  #16932

## Terminology used:

`Derived Output columns:` Output columns defined by each operator

`Output Columns:` The intersection of the operator's derived output columns and
the operator's parent's combined output pred columns. If the operator is the
root operator, its output columns would be the intersection of the operator's
derived output columns and the query's output columns.

`Combined Output Pred Columns:` A set containing operator's output columns and
its predicate's (if present) derived output columns

## Root Cause Analysis
-------------------------

Consider the below tree
```
+--A(Logical)
   |--B (Logical)
   |  |----C (Logical)
   |  |----D (Logical)
   |  +----E (Scalar)
   |
   +--F (Scalar)
```
In order to perform left join pruning in ORCA, one of the step is that each
operator(Logical) determines two sets. One set containing output columns of the
operator and second set containing combined output pred columns: For e.g. in
the above tree, operator A will create two sets. One set containing output
columns of A (let's call this set S1) and second set containing combined output
pred columns (let's call this set S2). In order to determine those two sets the
below algorithm is employed (Suppose we are determining for operator A)
1.) The operator A will find it's derived output columns and will take a
intersection with the query's output. The common values will be inserted into
set S1 and S2. The set S1 now contains the output columns of operator A.
2.)If a scalar operator is present for operator A then the derived output
columns of the scalar operator are also inserted into the set S2 (In this case
operator F's derived output columns)
3.) Now the set S2 will contain combined output pred columns for operator A.
4.) This set S2 is passed to the next function call in which the output columns
of operator B will be determined.

Now consider the below query:

**Setup:**
```
create table t1(a int primary key, b int) ;
create table t2(a int primary key, b int) ;
```

**Query:**

```
explain select t2.b from t1 left join t2 on t1.a = t2.a union all select t2.b
from t1 left join t2 on t1.a = t2.a;
```

**Algebrized tree:**

```
+--CLogicalUnionAll Output: ("b" (10)), Input: [("b" (10)), ("b" (28))]
   |--CLogicalLeftOuterJoin
   |  |--CLogicalGet "t1" ("t1"), Columns: ["a" (0), "b" (1)]
   |  |--CLogicalGet "t2" ("t2"), Columns: ["a" (9), "b" (10)]
   |  +--CScalarCmp (=)
   |     |--CScalarIdent "a" (0)
   |     +--CScalarIdent "a" (9)
   +--CLogicalLeftOuterJoin
      |--CLogicalGet "t1" ("t1"), Columns: ["a" (18), "b" (19)]
      |--CLogicalGet "t2" ("t2"), Columns: ["a" (27), "b" (28)]
      +--CScalarCmp (=)
         |--CScalarIdent "a" (18)
         +--CScalarIdent "a" (27)
```

Based on the above algorithm mentioned
The derived output columns of CLogicalSetOp
(CLogicalUnion/UnionAll/Intersect/IntersectAll/Except/ExceptAll) are from its
outer child. Based on the algorithm mentioned above, CLogicalUnionAll will
create two sets. One set containing output columns of itself and second set
containing combined output pred columns . The first set will contain b(10) and
the second set will also contain b(10). This second set will be passed to the
outer and inner CLogicalLeftOuterJoin operator for determining there output
columns. The outer CLogicalLeftOuterJoin will find its derived output columns
and will intersect it with the set passed by CLogicalUnionAll containing b(10).
After intersection we will get b(10) as the output column for outer
CLogicalLeftOuterJoin.  Similarly the inner CLogicalLeftOuterJoin will find its
derived output columns and will intersect it with the set passed by
CLogicalUnionAll containing b(10). But here the intersection will lead to a
empty output column set for inner CLogicalLeftOuterJoin. Now for pruning a left
join in a query, one of the condition that ORCA checks is that all output
columns of CLogicalLeftOuterJoin should only belong from the outer relation.
The method performing this check doesn't handle the case when the output
columns set is empty. Due to this all the checks pass in that method and ORCA
proceeds to prune the Join. So basically the issue is that ORCA is not able to
properly determine the output columns for the child operators when the parent
is CLogicalSetOp which is leading to unwanted pruning of joins. A similar issue
was fixed in PR #16690 which ensured that the combined output pred columns of
CLogicalSetOp is not empty because an empty set was leading to unwanted pruning
of left join. But that fix was not complete. It was incorrectly determining
combined output pred columns  of CLogicalSetOp which lead to #16932.

## Fix
--------
Updating the logic for determining the combined output pred columns  of
CLogicalSetOp. Since in the case of CLogicalSetOp, the output columns are from
the outer child, the inner child will always have empty output columns which is
not completely correct. So we need to also include the inner child columns in
the combined output pred columns  for CLogicalSetOP.