Enable local index by default and remove the else code
The file name is previously generated by uuid, which means
the data file is unordered. The ctid is hard to define.
The ctid used by DELETE/UPDATE is built temporarily on dynamic
shared memory.
However, manipulating dynamic shared memory is unfriendly to
parallel scan, because we use the pg functions to operate them.
We'll no longer maintain the old behavior that uses uuid as
file name, so the file name will become the block number as
part of the ctid.