Feature: PAX support visibility map for delete/update
PAX will rewrite the whole data file if any tuple in the data
file will be deleted/updated. The benefit is that all tuples
in the data file are visible, the column values are continuous.
The side effect amplifies write overhead and will slow scanning
tuples a bit.

The major bad side is that rewriting the whole data file will
bloat the index file. Using visibility map could reduce
the overhead for write and be friendly for index.