Modify Orca unit test asserts to run in retail build as well and fix unused variable warnings (#17307)
On my Mac, ninja fails when building in retail mode due to variables
that are only used in debug build.

Currently, many of our unit tests run with GPOS_ASSERT, which only runs
in assert (debug) builds. To make retail vs debug test builds more
consistent, this commit introduces "GPOS_UNITTEST_ASSERT" and a couple
other variations in assert.h, which are just a macro for
GPOS_RTL_ASSERT. We have this as a separate macro since we want to be
mindful of including retail asserts in production code. This gets
rid of the unused variable warnings/errors we were getting in various
compilers/environments, and should help reduce these going forward.