PAX: Add bitpacked column for arrow boolean type
In arrow columnar format, the boolean type is stored as bit-packed type.
When we return the record batch, we need to convert the pg storage format
to the arrow storage format. For further optimization, we need to store
the boolean type as a bit-packed type in the implementation of vectorized
direct storage.