Pax: storage format porc_vec support big group
After vectorization changes the maximum of record batch size, the porc_vec format on
PAX can also support large groups.

The reason why porc_vec format did not support large groups is that the single group
data of porc_vec cannot be split in vectorization path. PAX cannot do padding in the
middle of the data to ensure that the record batch buffers are always aligned with 8.