PAX: Fix visibility map to build ctid for vectorization
The offset in ctid is counted from 0 in a data file. It's calculated by
group_base_offset + range_begin + i where group_base_offset is the tuple
offset of the first tuple of the group in data file, range_begin is the
offset position of the group, and i is the i-th tuple in the current batch.

Expose the pax regression in Makefile, so we could easily run regression
test for PAX.