PAX: optimize min-max memory state
removed the OrcColumnStatsData, replace it with the MicroPartitionStatsData.
OrcColumnStatsData used to store the group min/max stats, But it should be consistent with the MicroPartitionStatsData.

After the current changes, we only need to maintain the structure MicroPartitionStatsData.

And in the current changes, the updated datum is no longer serialized into PB every update, but written into PB when Serialize is called.