PAX: add more exception message
The exceptions in PAX was simple, relying only on the exceptions type to prompt information.

But when debugging is needed, we often need some accompanying contextual information to quickly check the current error problem.

Therefore, in the current changes, exception is strengthened and all the contextual information we need added in exception.