PAX: add existexttoast to summary callback
In the previous PAX toast design, we uniformly named the toast files generated by PAX to
<blockname>.toast. This allows PAX to check whether the toast file exists or not by 
toast_file->exist().

Also the implementation of pax toast is unaware of storage_am/pg.

But after PAX was connected to gopher, the exist() interface has become unavailable.
Because if current file not exist, then gopher will do the throw -> try...catch -> 
return false which is very inefficient.

This forces PAX to record the toast_exist field(in the aux table) to avoid calling the
exist() interface.