PAX: remove the opfamily and build scan will not use the BRINAM
In pax filter, after we obtain the right value type from expr, we will try
to construct a scankey through the brinindex method. The scan key saves the
funcid/left type/right type which will used to do call the min/max functions.

In fact, building scankey through brinindex method may be incomplete. For example,
the bool type cannot construct the scankey.

In the current changes, brinindex is removed and we only verified the opfuncid(in expr)
exist in pg_operator.