PAX: disable numeric in bloom filter
Disable the numeric when PAX use the bloom filter do the filter.

Bloom filter will got wrong result in this case:

1. create table t(v1 numeric(20, 10)) using pax with(bloomfilter_columns='v1');
2. insert into t values(...);
3. select * from in_test_t where v1 in (2.0, 4.0);

Because there is `n_header` field in numeric type, in PG, different `n_header` will be considered to be the same numeric type.

So in wrong case, `2.0` will be the `numeric(1,0)`, but we store the `numeric(20,10)` results in the bloom filter.