PAX: don't detoast unread and short header datum
PAX will detoast the incoming datum(in tupleslot) by default. This is to prevent  write a PG toast datum.

But detoast will also cause the 2B header datum to become a 4B header datum.

If the current commit determines that it is a 2B header datum, it will no longer be detoast.