PAX:Filter support CAST in PG path
In the previous PR, PAX implemented a new sparse
filter. 

In the old version of sparse filter, the sparse
filter will skip the cast node directly or
directly call the comparison method for `opexpr`
with different left and right value types.
In current PR, added the `CastNode` to the new
implementation of the sparse filter to make it
more complete.