ORCA dsisable the update and delete on partitioned tables
After CBDB merged with PG14, we need to provide additional
support for the case where the update/delete is a dynamic (any)
scan operator. This is because in `ExecModifyTable`, PG14 has
an assumption: the current `resultRelInfo` will not be a root
partition table.

The current commit disables the update/delete generated by ORCA
on the partition table, and this behavior will be supported in
subsequent changes.