ORCA ignores empty or unsupported ext stats
STATS_EXT_DEPENDENCIES can be empty after analyzed. This is
because when the degree of all combinations is 0, PG will
not record the dependencies data in pg_statistic_ext_data.

The current changes ensure that ORCA will not fail to
generate a plan due to failure to read dependencies data,
and also ignore unsupported ext stats type(STATS_EXT_EXPRESSIONS).