Pax: fix io error in pax table when transaction abort
In an insert transaction, after obtaining a block_id from pax_fastsequence,
the block_id file and block_id.toast file will be created when file_system
->Open() calls are successful. When the transaction aborts, since pax_fastsequence
is implemented as update-in-place, whether the block_id increases depends on
whether the data has been flushed to disk. As the transaction aborts and xlog
won't be flushed, the next allocation may either reuse the current block_id
or use the next one.

Due to this non-deterministic behavior, the data file and toast file created
earlier may become orphaned files when the block_id is reused.

Therefore, O_CREAT | O_TRUNC is specified: if the file does not exist,
it will be created; if it exists, its previous content will be truncated
to avoid orphaned files.