[ORCA] Support PAX AM in ORCA
ORCA can generate plans for PAX(Access method) because PAX is registered in `RetrieveRelStorageType`,
but PAX will be registered as `ErelstorageAppendOnlyCols`.

Use the `ao version` as an example, if we still keep the `get_ao_version` logic (currently deleted),
then PAX will not be able to generate `IndexOnlyScan`, cause it have have registered in
`pg_appendonly`(should not and won't).

The current commit adds `ErelstoragePAX` storage type, not only to identify the PAX in DXL/Expression,
but also to do some preparatory work for optimization of PAX. For example, PAX can use its own
statistics as AM callbacks which can help ORCA better calculate the costs of `SeqScan`.