PAX: change length array to offset array
In PAX `PORC` format, length streaming is used to record the length of each
DATUM in a non-fixed-length column. The composition of length streaming is
equivalent to the length array, and its size equal to the number of rows.
When reading thenon-fixed-length column, PAX needs to use the length array
to calculate the offset array in advance, the offset array can help the
format reader quickly locate the middle row.

In commit "Performance/improve pax insert performance", PAX no longer builds
offsets array during the write phase, which actually breaks the assumption of
column: only some column methods distinguish between read and write paths.

In the current commit, the length streaming of PAX is changed to the offset streaming.

- On the read path, non-fixed-length column no need to build the offset array
- On the read path, only using the offset array is more cache-friendly
- On the write path, only the offset array needs to be built. And the performance
  is comparable to building the length array

Offset streaming also has disadvantages: the compression rate of offset streaming
is likely to be lower than that of length streaming. Currently, PAX does not support
DELTA encoding. Once DELTA encoding is supported, this disadvantages may can resolved.