ORCA: Pruning the output columns in CTE
In previous commits, we addressed the issue of "column
references not used by producer or consumer." However,
in practical cases, if a producer's output column is
used as a GROUP BY key, JOIN key, etc., by downstream
operators (within CTEs), the corresponding column
can not be pruned.

This commit adopts a new approach to resolve the issue:

1. We mark columns used in the consumer(Referenced by
upper-level operators) and notify the corresponding
producer.
2. The producer no longer marks its own output columns
as used. Instead, it finalizes its output columns
(i.e., the union of all columns required by consumers).
3. Since `NodeSharedScan` does not perform column projection,
we align the output columns in consumers (which are
guaranteed to be a subset of the producer’s output columns)
after the producer finalizes its output.

Note: Even if NodeSharedScan supported column projection,
it would not improve performance but complicate certain
cases. Thus, column projection was intentionally not
implemented in `NodeSharedScan`.