ORCA: deduplicate superset in AndOp when scalar condition has subset
This commit addresses scenarios where certain WHERE conditions
are subsets of others. For example, in WHERE A AND B, if condition
A is a superset of B, we can direct prune the condition A.

For most queries, eliminating redundant WHERE conditions to avoids
unnecessary inner joins when subqueries exist in the WHERE clause.
Due to the complexity of SQL, we can only target some simple cases
as superset. And only when subqueries exist, deduplication can
achieve actual performance improvement.

In this commit, the selection for superset is: subqueries without
`GROUP BY, WHERE, ORDER BY, LIMIT` as r-value. And the selection
for subset is: WHERE conditions containing the same l-value as
superset (or superset's l-value as joinkey with swapped values)
and the subquery with `GROUP BY, WHERE, ORDER BY, LIMIT` as l-values.
Of course, if the same supersets exist, it will also be pruned.

Example:

```
the schema: t1(v1, v2), t2(v3, v4)

The query:
    select * from t1,t2
      where
    	t1.v1 in (select v3 from t2)
    	and
    	t1.v1 in (select v3 from t2 where v3 < 100);

equivalent to:
    select * from t1,t2
      where
        t1.v1 t1.v1 in (select v3 from t2 where v3 < 100);
```