ORCA: Make partial agg use the ndistinct generated by segment to calculate the output rows
A new statistic STATISTIC_KIND_NDV_BY_SEGMENTS(hereinafter referred to as NDVOnSeg)
was added in the previous commit and is applied to ORCA in the current commit.

Due to the limited slots of statskind, not all cases have NDVOnSeg statistics.
However, if the current colstats has NDVOnSeg, NDVOnSeg will be used first,
otherwise, ORCA will fall back to using the histogram to calculate NDV (the
calculation result is close to the NDV in pg_statistic).

In fact, I think this commit is quite safe, because I only changed the NDV value
in `CColStats`, which means that the cost calculation will be more accurate when
the GbAgg operator is closer to the Scan operator.